% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfGrapher.R
\name{read_depth_per_position}
\alias{read_depth_per_position}
\title{Read depth per genomic position}
\usage{
read_depth_per_position(vcf_name)
}
\arguments{
\item{vcf_name}{The name of your vcf file}
}
\value{
A plot for each chromosome with the position of the SNP in base pair on the x axis and the average number of reads on the y axis.
}
\description{
Plotting read depth per genomic position
}
