% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfGrapher.R
\name{SNP_heterosigousity_histogram}
\alias{SNP_heterosigousity_histogram}
\title{Histogram of the proportion of heterozigousity per SNP}
\usage{
SNP_heterosigousity_histogram(vcf_name)
}
\arguments{
\item{vcf_name}{The name of your vcf file}
}
\value{
An histogram of heterosigousity ratio per SNP
#' @export
}
\description{
Plotting an histogram for the proportion of heterosigousity per SNP
}
