% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfGrapher.R
\name{SNP_heterosigousity_per_position}
\alias{SNP_heterosigousity_per_position}
\title{Heterozigousity per genomic position}
\usage{
SNP_heterosigousity_per_position(vcf_name)
}
\arguments{
\item{vcf_name}{The name of your vcf file}
}
\value{
A plot of heterosigousity ratio per SNP as a function of the position on the genome
#' @export
}
\description{
Plotting the proportion of heterosigousity per SNP on the y axis as a function of the genomic position of the SNP
}
